Do not style the pagination in the back end preview (see #9136)
Description
-----------

Currently the pagination for e.g. a gallery content element is also styled as if it was a regular back end navigation (for a DCA). This is a little confusing (especially in the `restore` view) and was not the case prior to the new pagination template in Contao 5 (it was always the case prior to Contao 5 though). This PR fixes that by only adding the `.pagination` class, if it is rendered in the front end.

Commits
-------

786c7524 Do not style pagination in back end
60bb9980 add default
4161f28b use scope instead
8bda098c merge with 5.x
51b00e59 Revert "use scope instead"
65ed01d3 use tl_pagination instead
24560252 merge with 5.x